refactor: reuse deepMerge & readJsonFile utils
