+++ packages/app-shell-vite-plugin/src/locales-utils.ts
@@ -1,35 +1,9 @@
 import fs from "node:fs";
 import path from "node:path";
 
-export const SUPPORTED_LOCALES_FILE = "supported-locales.json";
+import { deepMerge, readJsonFile } from "./utils.js";
 
-/**
- * Deep merge two objects. `override` keys take priority over `base`.
- */
-export function deepMerge(
-  base: Record<string, unknown>,
-  override: Record<string, unknown>,
-): Record<string, unknown> {
-  const result = { ...base };
-  for (const [key, value] of Object.entries(override)) {
-    if (
-      value != null &&
-      typeof value === "object" &&
-      !Array.isArray(value) &&
-      result[key] != null &&
-      typeof result[key] === "object" &&
-      !Array.isArray(result[key])
-    ) {
-      result[key] = deepMerge(
-        result[key] as Record<string, unknown>,
-        value as Record<string, unknown>,
-      );
-    } else {
-      result[key] = value;
-    }
-  }
-  return result;
-}
+export const SUPPORTED_LOCALES_FILE = "supported-locales.json";
 
 /**
  * Reads a `supported-locales.json` file and returns its contents as an array
@@ -39,7 +13,7 @@ export function deepMerge(
 export function readSupportedLocales(filePath: string): string[] {
   if (!fs.existsSync(filePath)) return [];
   try {
-    const data: unknown = JSON.parse(fs.readFileSync(filePath, "utf-8"));
+    const data = readJsonFile(filePath);
     if (Array.isArray(data) && data.every((item) => typeof item === "string")) {
       return data;
     }
@@ -123,22 +97,6 @@ export function computeSupportedLocales(
   }
 
   return result.toSorted();
-}
-
-/**
- * Parses a JSON file, wrapping parse errors with the file path for
- * actionable diagnostics.
- */
-export function readJsonFile(filePath: string): unknown {
-  const raw = fs.readFileSync(filePath, "utf-8");
-  try {
-    return JSON.parse(raw);
-  } catch (cause) {
-    throw new Error(
-      `[app-shell-locales] Failed to parse ${filePath}: ${cause instanceof Error ? cause.message : cause}`,
-      { cause },
-    );
-  }
 }
 
 /**
